+++ clikt-mordant/src/commonTest/kotlin/com/github/ajalt/clikt/parsers/AtFileTest.kt
@@ -157,7 +157,7 @@ class AtFileTest {
         class C : TestCommand() {
             init {
                 context {
-                    expandArgumentFiles = false
+                    argumentFileReader = null
                 }
             }